draw: Introduce a dummy draw function
This function only increments the block offsets, and does not draw
anything.

Though this approach duplicates code and can be considered somewhat
ugly, it increases the speed of calc_next_mask by several orders of
magnitude, which will become useful for pre-indexing a file, or seeking,
in the future.

A massive speedup is gained in the find-next-different functionality.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>